+++ dash_pydantic_form/utils.py
@@ -139,7 +139,7 @@ def get_model_value(item: BaseModel, field: str, parent: str, allow_default: boo
         return subitem[field]
     except:
         if allow_default:
-            subitem_cls = get_subitem_cls(item.__class__, parent)
+            subitem_cls = get_subitem_cls(item.__class__, parent, item=item)
             if not is_subclass(subitem_cls, BaseModel):
                 return None
             field_info = subitem_cls.model_fields[field]
@@ -183,7 +183,7 @@ def is_idx_template(val: str):
     return bool(re.findall(r"^\{\{[\w|\{\}]+\}\}$", val))
 
 
-def get_subitem_cls(model: type[BaseModel], parent: str) -> type[BaseModel]:
+def get_subitem_cls(model: type[BaseModel], parent: str, item: BaseModel | None = None) -> type[BaseModel]:
     """Get the subitem class of a model at a given parent.
 
     e.g., get_subitem_cls(Person, "au_metadata") = AUMetadata
@@ -204,26 +204,38 @@ def get_subitem_cls(model: type[BaseModel], parent: str) -> type[BaseModel]:
     if isinstance(second_part, str) and second_part.isdigit():
         second_part = int(second_part)
 
-    first_annotation = get_non_null_annotation(model.model_fields[first_part].annotation)
+    field_info = model.model_fields[first_part]
+    first_annotation = get_non_null_annotation(field_info.annotation)
+    subitem = get_subitem(item, first_part) if item is not None else None
+    if Type.classify(first_annotation, field_info.discriminator) == Type.DISCRIMINATED_MODEL:
+        if not item:
+            raise TypeError("Discriminated models with nesting need passing item data to be displayed")
+        discriminator_value = None if subitem is None else getattr(subitem, field_info.discriminator, None)
+        subitem_cls, _ = handle_discriminated(
+            item.__class__, path, first_annotation, field_info.discriminator, discriminator_value
+        )
+        return get_subitem_cls(subitem_cls, SEP.join(path[1:]), item=subitem)
     if (
-        get_origin(first_annotation) == list
+        get_origin(first_annotation) is list
         and (isinstance(second_part, int) or is_idx_template(second_part))
         and get_args(first_annotation)
     ):
         return get_subitem_cls(
             get_non_null_annotation(get_args(first_annotation)[0]),
             SEP.join(path[2:]),
+            item=subitem,
         )
     if (
-        get_origin(first_annotation) == dict
+        get_origin(first_annotation) is dict
         and (isinstance(second_part, int) or is_idx_template(second_part))
         and get_args(first_annotation)
     ):
         return get_subitem_cls(
             get_non_null_annotation(get_args(first_annotation)[1]),
             SEP.join(path[2:]),
+            item=subitem,
         )
-    return get_subitem_cls(first_annotation, SEP.join(path[1:]))
+    return get_subitem_cls(first_annotation, SEP.join(path[1:]), item=subitem)
 
 
 def handle_discriminated(model: type[BaseModel], parent: str, annotation: type, disc_field: str, disc_val: Any):